Switch to pnpm, fix eslint and tsconfig, temporarily remove workspaces
To simplify, I have removed the electorn backend and network-utils
workspaces, since they aren't yet being used anywhere. This reduced
the dependency tree to allow me to understand it better.

Looking into issues with storybook and encountered a lot of issues with
eslint, mainly to do with peerDependencies not being installed. Switched
to pnpm which allowed me to track down and install the missing deps.

This in turn allowed me to realize that the UI module was not importing
the TS config.