Fix issues to do with stale data in the interview route
It seems that the interview/new route was getting stale data from the
allowAnonymousRecruitment query no matter what I tried.

I tracked this down to the implementation of the unstable nextCache
link that we were using for server side tRPS. Some investigation shows
that it doesn't allow for a revalidation time less than 1 second, which
is directly at odds with the Next docs which say to pass `revalidate: 0`
to disable caching.

So, here I switch to using a HTTP/fetch based link for server calls,
which has actually had a number of additional benefits:

- We can pass `{ context: { revalidate: 0 } }` as options to any
  server call to disable caching. This is a much more explicit
  way of doing it than the unstable `nextCache` link and using the
  `helpers` from the proxy method.
- I was able to turn on batching for server tRPC calls, which is
  hopefully going to improve performance. The app feels wayyy faster,
  but it could be my imagination.
- I was able to remove the proxy client entirely, and remove the use
  of the `helpers` from the session router. The redirects are all still
  working perfectly, and the app is still able to get the session
  server-side from what I can tell. I _hope_ this might help with the
  deployment issue.