Fix for promptId related bugs
This fixes bugs related to the lack of a promptId in certain contexts.

The issue was caused by faulty selectors that assumed promptIndex (not
ID!) was being passed into the stage component as a prop. I removed this
during refactoring as it would cause the stage to re-render when the
prompt changed (so was incorrect).

The new implementation gets the promptId from the session, via the
promptIndex. This is implemented within the `getPromptIndex` session
selector.